refactor: change controllers naming
